Moved AuthemnticateUserTest to GeneralServiceTest
